Organize folders for WebSockets feature
