+++ test/Ocelot.UnitTests/Middleware/OcelotPipelineExtensionsTests.cs
@@ -8,7 +8,7 @@ using Ocelot.DownstreamUrlCreator.Middleware;
 using Ocelot.LoadBalancer.Middleware;
 using Ocelot.Middleware;
 using Ocelot.Request.Middleware;
-using Ocelot.WebSockets.Middleware;
+using Ocelot.WebSockets;
 
 namespace Ocelot.UnitTests.Middleware
 {
